[ADD] validation 관련 annotation 추가
