--- src/main/kotlin/com/soom/monolithic_api/domain/account/signup/request/TeacherSignupRequest.kt
@@ -1,15 +1,0 @@
-package com.soom.monolithic_api.domain.account.signup.request
-
-import com.soom.monolithic_api.domain.account.common.data.type.GenderType
-import com.soom.monolithic_api.domain.account.common.data.type.TeacherType
-import java.time.LocalDate
-
-data class TeacherSignupRequest(
-    override val emailToken: String,
-    override val password: String,
-    override val name: String,
-    override val gender: GenderType,
-    override val birth: LocalDate,
-    val major: String,
-    val teacherType: TeacherType
-): SignupRequest(emailToken, password, name, gender, birth)
